+++ st-tdg-test/src/test/java/test4giis/tdrules/tdg/st/test/market/TestMarketFuncCarts.java
@@ -6,8 +6,10 @@ import giis.tdrules.store.loader.DataLoader;
 import giis.tdrules.store.loader.IAttrGen;
 
 /**
- * Tests de funcionalidad de Carts
- * Generación: QAGrow y claves se realiza en el backend
+ * Carts Functionality Tests.
+ * - Use the live SUT where the uids of each generated object are generated by the backend.
+ * - Use the dictionary
+ * - Include Users, Carts and Orders
  */
 public class TestMarketFuncCarts extends BaseMarket {
 	
@@ -18,6 +20,7 @@ public class TestMarketFuncCarts extends BaseMarket {
 	
 	@Test
 	public void testDictUserDTOResByName() {
+		// users
 		String query = queryUserByName;
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -25,29 +28,20 @@ public class TestMarketFuncCarts extends BaseMarket {
 		assertData("func-UsersByName.txt", dg);
 	}
 
-	/*
-	 * Inserta carts pero no items ni productos. Para cubrir las reglas de cobertura no son necesarios
-	 * Son carts vacíos. 
-	 */
 	@Test
 	public void testDictCartDTOByUser() {
-		String query = "tds CartDTO where user='lucia@email.com'";
+		// empty carts
+		String query = queryCartByUser;
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
 		generateAndLoad(dg, query, dict);
 		assertData("func-CartsByUser.txt", dg);
 	}
 	
-	/* Inserta carts y sus items.
-	 * Para que se inserte un producto en un carrito, debe estar disponible, necesario incluirlo en la query.
-	 * Si no se indica en la query, se genera por DataGenerator y es no disponible (available = false),
-	 * entonces al intentar añadirlo via api, no se inserta en la tabla (lo que es correcto)
-	 * Hay otra regla de cobertura que no se cubre con una única base de datos:
-	 * - user='lucia...' and productId=1 and quantity != 5
-	 * En este caso, desde QAGrow ya no se generan los datos.
-	 */
 	@Test
 	public void testDictCartDTOByUserProductQuantity() {
+		// carts and items
+		// products must be available
 		String query = queryCartByUserProductQuantity;
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -55,14 +49,13 @@ public class TestMarketFuncCarts extends BaseMarket {
 		assertData("func-CartsByUserProductQuantity.txt", dg);
 	}
 	
-	/* Inserta order a un usuario.
-	 * Para que cree una order, el carrito no puede estar vacio.
-	 * Por tanto, habrá una query1 para generar carritos con productos y query2 para generar la orden del carrito
-	 */
 	@Test
 	public void testOrderDTOByUser() {
+		// order (cart must not be empty)
+		// query 1 generates a non-empty cart, 
+		// query 2 generates the order
 		String query1= queryCartByUserProductQuantity;
-		String query2 = "tds OrderDTO where userAccount='pepe@email.com' ";
+		String query2 = queryOrderByUser;
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
 		generateAndLoad(dg, new String[] {query1, query2}, dict);
