+++ st-tdg-test/src/test/java/test4giis/tdrules/tdg/st/test/market/TestMarketFuncProducts.java
@@ -5,8 +5,10 @@ import giis.tdrules.store.loader.DataLoader;
 import giis.tdrules.store.loader.IAttrGen;
 
 /**
- * Tests de funcionalidad de Products
- * Generación: QAGrow y claves se realiza en el backend
+ * Products Functionality Tests
+ * - Use the live SUT where the uids of each generated object are generated by the backend.
+ * - Use the dictionary
+ * - Include Products, Distilleries and Regions
  */
 public class TestMarketFuncProducts extends BaseMarket {
 
@@ -15,11 +17,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 		return true;
 	}
 	
-	/**
-	 * Todos los productos de una destileria
-	 */
 	@Test
 	public void testProductsByDistillery() {
+		// products of a distillery
 		String query =  "tds ProductDTOReq where distillery ='Ardbeg'";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -27,12 +27,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 		assertData("func-ProductsByDistillery.txt", dg);
 	}
 
-	
-	/**
-	 * Productos de una destileria disponibles
-	 */
 	@Test
 	public void testProductsByDistilleryAvaliable() {
+		// available products of a distillery 
 		String query =  "tds ProductDTOReq where distillery ='Ardbeg' and available = 1";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -40,11 +37,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 		assertData("func-ProductsByDistilleryAvailable.txt", dg);
 	}
 	
-	/**
-	 * Todos los productos de una destileria no disponibles
-	 */
 	@Test
 	public void testProductsByDistilleryNotAvailable() {
+		// non-available products of a distillery
 		String query =  "tds ProductDTOReq where distillery ='Ardbeg' and available = 0";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -52,11 +47,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 		assertData("func-ProductsByDistilleryNotAvailable.txt", dg);
 	}
 	
-	/**
-	 * Productos de una destileria disponibles entre dos precios
-	 */
 	@Test
 	public void testProductsByDistilleryPrice() {
+		// available products of a distillery, price between two values
 		String query =  "tds ProductDTOReq where distillery ='Ardbeg' and price < 100 and price > 5";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -64,11 +57,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 		assertData("func-ProductsByDistilleryPrice.txt", dg);
 	}
 	
-	/**
-	 * Productos de una destileria disponibles entre dos precios
-	 */
 	@Test
 	public void testProductsByDistilleryPriceAge() {
+		// available products of a distillery and an age, price between two values
 		String query =  "tds ProductDTOReq where distillery ='Ardbeg' and price < 100 and price > 5 and age = 12";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -77,11 +68,9 @@ public class TestMarketFuncProducts extends BaseMarket {
 	}
 	
 	
-	/**
-	 * Prueba de generacion de distillery y region con claves en el backend, fijando distillery.title
-	 */
 	@Test
 	public void testDistilleryByTitle() {
+		// distillery 
 		String query =  "tds DistilleryDTOReq where title ='Ardbeg'";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
@@ -89,16 +78,24 @@ public class TestMarketFuncProducts extends BaseMarket {
 		assertData("func-DistilleryByTitle.txt", dg);
 	}
 	
-	/**
-	 * Prueba de generacion de distillery y region con claves en el backend, fijando distillery.region (region es maestro de distillery)
-	 */
 	@Test
 	public void testDistilleryByRegion() {
+		// distillery of a region
 		String query =  "tds DistilleryDTOReq where region ='Islay'";
 		IAttrGen dict=getDictionaryAttrGen();
 		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
 		generateAndLoad(dg, query, dict);
 		assertData("func-DistilleryByRegion.txt", dg);
+	}
+	
+	@Test
+	public void testRegionByName() {
+		// region
+		String query =  "tds RegionDTOReq where name ='Campbeltown'";
+		IAttrGen dict=getDictionaryAttrGen();
+		DataLoader dg = getLiveDataLoader().setAttrGen(dict);
+		generateAndLoad(dg, query, dict);
+		assertData("func-RegionByName.txt", dg);
 	}
 
 }
