+++ st-tdg-test/src/test/java/test4giis/tdrules/tdg/st/test/market/TestMarketSchemaLocal.java
@@ -9,7 +9,8 @@ import giis.tdrules.model.io.TdSchemaXmlSerializer;
 import giis.tdrules.openapi.model.TdSchema;
 
 /**
- * Genera el esquema a partir de la especificacion (leida de archivo) y lo comprueba en formato json y xml
+ * Checks the transformation of the OpenApi specification into the TdSchema model
+ * (json, xml and Mermaid graph)
  */
 public class TestMarketSchemaLocal extends BaseMarket {
