--- st-tdg-test/src/test/resources/market/qagrow-livebackid-DistilleryByTitle.txt
@@ -1,3 +1,0 @@
-"distillery":{"id":1,"title":"Ardbeg","region":{"id":1,"name":"Campbeltown","subtitle":"5","description":"3","color":"2"},"description":"103"}
-"distillery":{"id":2,"title":"Balvenie","region":{"id":1,"name":"Campbeltown","subtitle":"5","description":"3","color":"2"},"description":"203"}
-"region":{"id":1,"name":"Campbeltown","subtitle":"5","description":"3","color":"2"}
